Resolve requested changes
